requirements.txt
sentence-transformers
faiss-cpu
openai
pyyaml
pandas
tqdm
python-dotenv
underthesea
rapidfuzz
numpy
